requirements.txt
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.0.0
Flask-SocketIO==5.5.1
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
h11==0.16.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Naked==0.1.32
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pyparsing==3.2.5
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.15.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
shellescape==3.8.1
simple-websocket==1.1.0
urllib3==2.6.1
Werkzeug==3.0.1
wsproto==1.3.2
